requirements.txt
streamlit
exa-py
python-dotenv
requests
